Here is a self-contained login page
it does not integrate with the website, but it is a standalone working part. Feel free to download it and give it a try (though i will just explain what it does below):
it does the stuff you would expect a jwt login to do, using mongoDB to minimally store refresh tokens for an admin user. 
Valid login username:password pairs are hard coded, but the purpose is to require a valid access token while the user stays on the website pages, especially the stuff that only admins should see.

I am not 100% sure what those things will look like, or how to implement it smoothly as of now, I am submitting mainly for the purpose of having some working part to see so progress can be evaluated concretely.